Another sane bit-width type cleanup.
git-svn-id: https://code.crt.realtors.org/svn/odbcretsdrv/ezrets/branches/build-64-bit@1406 fe7ac059-47f5-0310-8599-d0339b4e125c

+++ project/common/src/DataSource.cpp
@@ -949,7 +949,7 @@ void DataSource::SetFromOdbcConnectionString(string connectionString)
          i != tokens.end(); i++)
     {
         string tok = *i;
-        int epos = tok.find("=");
+        size_t epos = tok.find("=");
         string key = tok.substr(0, epos);
         string value = tok.substr(epos+1, tok.size());
